Reformat frontend directory structure
