--- frontend/src/Components/ForgotPassword/ForgotPassword.jsx
@@ -1,12 +1,0 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import '../LoginSignup/LoginSignup.css';  
-
-//Firebase functionality methods from the firebase.js file 
-import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '../../firebase';
-
-export default function ForgotPassword(){
-
-
-}
