+++ frontend/src/Components/Helpers/ForgotPassword/ForgotPassword.jsx
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import '../../Pages/LoginSignup/LoginSignup.css';  
+
+//Firebase functionality methods from the firebase.js file 
+import { signInWithEmailAndPassword } from 'firebase/auth';
+import { auth } from '../../../firebase';
+
+export default function ForgotPassword(){
+
+
+}
