+++ frontend/src/Components/Helpers/StructuredDiffEditor/ResumeDiffViewer.jsx
