--- frontend/src/Components/UploadPdf/PdfUtils.js
@@ -1,6 +1,0 @@
-//this will have utility functions for uploading the pdf like validating the pdf and size
-
-export function isValidPDF(file){
-    const maxSizeMB = 6; //maximum size of file is 6 megabites
-    //const validType = file type 
-}
